requirements.txt
streamlit==1.37.1
pypdf==4.3.1
sentence-transformers==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
pdfplumber==0.11.4
pdf2image==1.17.0
pytesseract==0.3.13
Pillow==10.4.0